refactor: move aigc task to model
